made PORT a constant

+++ components/common/networking/inc/ethernet.h
@@ -29,8 +29,8 @@
  *         RESULT_OK: if ethernet started properly
  */
 result_t ETH_init(linkstatus_callback_t link_state_change_callback,
-                           uint8_t ip[4], uint8_t netmask[4],
-                           uint8_t gateway[4], uint8_t mac_address[6]);
+                  uint8_t ip[4], uint8_t netmask[4], uint8_t gateway[4],
+                  uint8_t mac_address[6]);
 
 /**
  * @brief Initialize the UDP protocol control block.
@@ -73,7 +73,7 @@ void ETH_udp_init(uint8_t sender_prio_num, QueueHandle_t *send_queues,
  *      Priority level used for transmission. Must be less than the value
  *      specified in ETH_udp_init().
  */
-void ETH_udp_send(uint8_t ip[4], uint8_t port, uint8_t *payload,
+void ETH_udp_send(uint8_t ip[4], int port, uint8_t *payload,
                   uint16_t payload_len, uint8_t prio_num);
 
 /**
